change short to int, byte to int, java has no unsigned short ,  unsigned byte

+++ src/main/java/com/silence/vm/Memory.java
@@ -2,7 +2,7 @@ package com.silence.vm;
 
 public class Memory {
     public static final int MEMORY_MAX = 1 << 16;
-    private static final short[] memory = new short[MEMORY_MAX];
+    private static final int[] memory = new int[MEMORY_MAX];
     private static boolean keyPressing = false;
 
     /**
@@ -10,7 +10,7 @@ public class Memory {
      * We can’t read and write to the memory array directly, but must instead call setter and getter functions.
      * When memory is read from KBSR, the getter will check the keyboard and update both memory locations.
      */
-    public static short mem_read(int address){
+    public static int mem_read(int address){
         if(address == Registers.MR_KBSR){
             if(check_key()){
                 memory[Registers.MR_KBSR] = (short) (1 << 15);
@@ -26,7 +26,7 @@ public class Memory {
         return keyPressing;
     }
 
-    public static void mem_write(int address, short val){
+    public static void mem_write(int address, int val){
         memory[address] = val;
     }
